locally tested fixes, added tests
