+++ .github/workflows/staging.yml
@@ -74,8 +74,8 @@ jobs:
     # NOTE: could save this as a build artifact or put on website
     - name: Export Full Conda Environment
       run: |
-        docker run ${DOCKER_ORG}/${{ matrix.IMAGE }}:master conda env export
+        docker run ${DOCKER_ORG}/${{ matrix.IMAGE }}:master conda list --explicit
 
     - name: Test Image
       run: |
-        docker run -u root -v $PWD:/home/jovyan ${DOCKER_ORG}/${{ matrix.IMAGE }}:master ./run_tests.sh
+        docker run -v $PWD:/home/jovyan ${DOCKER_ORG}/${{ matrix.IMAGE }}:master ./run_tests.sh ${{ matrix.IMAGE }}
