+++ .gitignore
@@ -3,6 +3,7 @@
 .bash_history
 .config
 dask-worker-space
+.DS_Store
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
