+++ base-image/condarc
@@ -6,7 +6,8 @@ show_channel_urls: true
 # NOTE: defaults only work for `conda create -f`, not `conda env create -f`!
 # https://github.com/conda/conda/issues/9580
 create_default_packages:
+  - python=3.7
   - pangeo-notebook=0.0.2
   - ipykernel
-pinned_packages:
-  - pangeo-notebook=0.0.2
+#pinned_packages:
+#  - pangeo-notebook=0.0.2
