+++ base-image/init_conda.sh
@@ -1,0 +1,3 @@
+# activate pangeo environment on login
+. "/srv/conda/etc/profile.d/conda.sh"
+conda activate pangeo
