--- base-notebook/environment.yml
@@ -1,5 +1,0 @@
-name: pangeo
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
