--- base-notebook/packages.txt
@@ -1,4 +1,0 @@
-# This file may be used to create an environment using:
-# $ conda create --name <env> --file <this file>
-# platform: linux-64
-python=3.7.6
