--- base-notebook/postBuild
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-# labextensions
-jupyter labextension install --clean dask-labextension \
-                             @jupyter-widgets/jupyterlab-manager
