+++ base-notebook/readme.md
@@ -1,0 +1,1 @@
+Simply using a Dockerfile with `FROM pangeodev/base-image:[tag]` will result in a valid minimal base-notebook
