--- ml-notebook/environment.yml
@@ -1,45 +1,0 @@
-name: pangeo
-channels:
-  - conda-forge
-  - pangeo
-dependencies:
-  - python=3.7
-  - pangeo-notebook=0.0.2
-# Additional libraries
-  - cartopy
-  - cupy
-  - dask-ml
-  - datashader
-  - fsspec
-  - gcsfs
-  - geopandas
-  - geoviews-core
-  - h5netcdf
-  - h5py
-  - holoviews
-  - hvplot
-  - intake
-  - intake-esm
-  - intake-stac
-  - intake-xarray
-  - ipyleaflet
-  - jupyterlab-nvdashboard
-  - jupyter-panel-proxy
-#  - keras-gpu
-  - lz4
-  - matplotlib-base
-  - numpy
-  - pandas
-  - panel
-  - pytorch-cpu
-  - rasterio
-  - s3fs
-  - sat-search
-  - sat-stac
-  - satpy
-  - scikit-image
-  - scikit-learn
-  - scipy
-#  - tensorflow-gpu
-  - xarray
-  - zarr
