--- ml-notebook/postBuild
@@ -1,6 +1,0 @@
-#!/bin/bash
-jupyter labextension install --clean dask-labextension \
-                             @jupyter-widgets/jupyterlab-manager \
-                             @pyviz/jupyterlab_pyviz \
-                             jupyter-leaflet
-#EOF
