+++ pangeo-notebook/apt.txt
@@ -1,1 +1,2 @@
+git
 vim
