--- pangeo-notebook/environment.yml
@@ -1,41 +1,0 @@
-name: pangeo
-channels:
-  - conda-forge
-dependencies:
-  - python=3.7
-  - pangeo-notebook=0.0.2
-  # Additional libraries
-  - cartopy
-  - dask-ml
-  - datashader
-  - fsspec
-  - gcsfs
-  - geopandas
-  - geoviews-core
-  - h5netcdf
-  - h5py
-  - holoviews
-  - hvplot
-  - intake
-  - intake-esm=2019
-  - intake-stac
-  - intake-xarray
-  - ipyleaflet
-  - ipywidgets
-  - jupyter-panel-proxy
-  - lz4
-  - matplotlib-base
-  - nomkl
-  - numpy
-  - pandas
-  - panel
-  - rasterio
-  - s3fs
-  - sat-search
-  - sat-stac
-  - satpy
-  - scikit-image
-  - scikit-learn
-  - scipy
-  - xarray
-  - zarr
