+++ pangeo-notebook/packages.txt
@@ -1,0 +1,46 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: linux-64
+python=3.7.6
+cartopy
+dask-ml
+datashader
+fsspec
+gcsfs
+geopandas
+geoviews-core
+h5netcdf
+h5py
+holoviews
+hvplot
+intake
+intake-esm
+intake-geopandas
+intake-stac
+intake-xarray
+ipywidgets
+jupyter-panel-proxy
+lz4
+matplotlib-base
+metpy
+nomkl
+numpy
+numcodecs
+pandas
+panel
+pydap
+python-blosc
+rasterio
+rio-cogeo
+s3fs
+sat-search
+sat-stac
+satpy
+scikit-image
+scikit-learn
+scipy
+sparse
+xarray
+xesmf
+xpublish
+zarr
