+++ pangeo-notebook/postBuild
@@ -1,5 +1,7 @@
 #!/bin/bash
-jupyter labextension install --clean dask-labextension \
-                             @jupyter-widgets/jupyterlab-manager \
-                             @pyviz/jupyterlab_pyviz \
-                             jupyter-leaflet
+
+jupyter labextension install --clean \
+                            @pyviz/jupyterlab_pyviz \
+                            jupyter-leaflet
+                            
+#EOF
