+++ pangeo-notebook/requirements.txt
@@ -1,0 +1,1 @@
+# add pip packages here
